Update Deliverable 2
Couple of issues while I do have it printing in order of code the current problem is that I want it to be able to return back to a keyword value and then go right to make the parse tree correct however at this point in time it doesn't do that. The other problem is it doesn't print the very last value for whatever reason. However it does seem to currently print in order which is a major plus.